+++ electron-app/utils.js
@@ -1,5 +1,3 @@
-/* eslint-env browser */
-
 // Utils.js - Enhanced Utility exports for FitFileViewer Electron app
 /**
  * Exposes utility functions globally for use in index.html and other scripts.